+++ demo/AppMode.ts
@@ -2,19 +2,25 @@ import * as R from 'ramda';
 
 import Environment from 'core/environment';
 
-import { generateMockData, IDataMock } from './data';
+import { generateMockData, IDataMock, generateSpaceMockData } from './data';
 import {
     ContentStyle,
     PropsWithDefaults,
     ChangeAction,
     ChangeFailure,
-    IVisibleColumn
+    IVisibleColumn,
+    ColumnType
 } from 'dash-table/components/Table/props';
+import { TooltipSyntax } from 'dash-table/tooltips/props';
 
 export enum AppMode {
+    Date = 'date',
     Default = 'default',
+    FixedTooltips = 'fixed,tooltips',
     FixedVirtualized = 'fixed,virtualized',
     ReadOnly = 'readonly',
+    ColumnsInSpace = 'columnsInSpace',
+    Tooltips = 'tooltips',
     Typed = 'typed',
     Virtualized = 'virtualized'
 }
@@ -35,7 +41,6 @@ function getBaseTableProps(mock: IDataMock) {
             },
             editable_name: true,
             deletable: true
-            //     type: 'dropdown'
         })),
         column_static_dropdown: [
             {
@@ -62,11 +67,13 @@ function getBaseTableProps(mock: IDataMock) {
     };
 }
 
-function getDefaultState(): {
+function getDefaultState(
+    generateData: Function = generateMockData
+): {
     filter: string,
     tableProps: Partial<PropsWithDefaults>
 } {
-    const mock = generateMockData(5000);
+    const mock = generateData(5000);
 
     return {
         filter: '',
@@ -97,6 +104,67 @@ function getReadonlyState() {
     return state;
 }
 
+function getSpaceInColumn() {
+    const state = getDefaultState(generateSpaceMockData);
+    state.tableProps.filtering = true;
+
+    return state;
+}
+
+function getFixedTooltipsState() {
+    const state = getTooltipsState();
+
+    state.tableProps.n_fixed_columns = 3;
+    state.tableProps.n_fixed_rows = 4;
+
+    return state;
+}
+
+function getTooltipsState() {
+    const state = getDefaultState();
+
+    state.tableProps.tooltip_delay = 250;
+    state.tableProps.tooltip_duration = 1000;
+    state.tableProps.tooltips = {
+        ccc: [
+            { type: TooltipSyntax.Markdown, value: `### Go Proverb\nThe enemy's key point is yours` },
+            { type: TooltipSyntax.Markdown, value: `### Go Proverb\nPlay on the point of symmetry` },
+            { type: TooltipSyntax.Markdown, value: `### Go Proverb\nSente gains nothing` },
+            { type: TooltipSyntax.Text, value: `Beware of going back to patch up` },
+            { type: TooltipSyntax.Text, value: `When in doubt, Tenuki` },
+            `People in glass houses shouldn't throw stones`
+        ]
+    };
+    state.tableProps.column_static_tooltip = {
+        ccc: { type: TooltipSyntax.Text, value: `There is death in the hane` },
+        ddd: { type: TooltipSyntax.Markdown, value: `Hane, Cut, Placement` },
+        rows: `Learn the eyestealing tesuji`
+    };
+    state.tableProps.column_conditional_tooltips = [{
+        if: {
+            column_id: 'aaa-readonly',
+            filter: `aaa is prime`
+        },
+        type: TooltipSyntax.Markdown,
+        value: `### Go Proverbs\nCapture three to get an eye`
+    }, {
+        if: {
+            column_id: 'bbb-readonly',
+            row_index: 'odd'
+        },
+        type: TooltipSyntax.Markdown,
+        value: `### Go Proverbs\nSix die but eight live`
+    }, {
+        if: {
+            column_id: 'bbb-readonly'
+        },
+        type: TooltipSyntax.Markdown,
+        value: `### Go Proverbs\nUrgent points before big points\n![Sensei](https://senseis.xmp.net/images/stone-hello.png)`
+    }];
+
+    return state;
+}
+
 function getTypedState() {
     const state = getDefaultState();
 
@@ -106,6 +174,37 @@ function getTypedState() {
             failure: ChangeFailure.Reject
         };
     }, state.tableProps.columns || []);
+
+    return state;
+}
+
+function getDateState() {
+    const state = getTypedState();
+
+    (state.tableProps.columns || []).forEach(column => {
+        if (column.id === 'ccc') {
+            column.name = ['Date', 'only'];
+            column.type = ColumnType.Datetime;
+            column.validation = { allow_YY: true };
+            (state.tableProps.data || []).forEach((row, i) => {
+                const d = new Date(Date.UTC(2018, 0, 1));
+                // three day increment
+                d.setUTCDate(3 * i + 1);
+                // date only
+                row.ccc = d.toISOString().substr(0, 10);
+            });
+        } else if (column.id === 'ddd') {
+            column.name = ['Date', 'with', 'time'];
+            column.type = ColumnType.Datetime;
+            (state.tableProps.data || []).forEach((row, i) => {
+                const d = new Date(Date.UTC(2018, 0, 1));
+                // two hours and 11 seconds increment
+                d.setUTCSeconds(i * 7211);
+                // datetime ending with seconds
+                row.ddd = d.toISOString().substr(0, 19).replace('T', ' ');
+            });
+        }
+    });
 
     return state;
 }
@@ -152,10 +251,18 @@ function getState() {
     const mode = Environment.searchParams.get('mode');
 
     switch (mode) {
+        case AppMode.Date:
+            return getDateState();
+        case AppMode.FixedTooltips:
+            return getFixedTooltipsState();
         case AppMode.FixedVirtualized:
             return getFixedVirtualizedState();
         case AppMode.ReadOnly:
             return getReadonlyState();
+        case AppMode.ColumnsInSpace:
+            return getSpaceInColumn();
+        case AppMode.Tooltips:
+            return getTooltipsState();
         case AppMode.Virtualized:
             return getVirtualizedState();
         case AppMode.Typed: