+++ src/dash-table/derived/style/props.ts
@@ -1,30 +1,11 @@
-import { ColumnId, ColumnType } from 'dash-table/components/Table/props';
+import {
+    ConditionalBasicFilter,
+    ConditionalDataCell,
+    ConditionalCell,
+    ConditionalHeader
+} from 'dash-table/conditional';
+
 import IStyle from './IStyle';
-
-export interface IConditionalElement {
-    filter?: string;
-}
-
-export interface IIndexedHeaderElement {
-    header_index?: number | 'odd' | 'even';
-}
-
-export interface IIndexedRowElement {
-    row_index?: number | 'odd' | 'even';
-}
-
-export interface INamedElement {
-    column_id?: ColumnId;
-}
-
-export interface ITypedElement {
-    column_type?: ColumnType;
-}
-
-type ConditionalBasicFilter = INamedElement & ITypedElement;
-type ConditionalDataCell = IConditionalElement & IIndexedRowElement & INamedElement & ITypedElement;
-type ConditionalCell = INamedElement & ITypedElement;
-type ConditionalHeader = IIndexedHeaderElement & INamedElement & ITypedElement;
 
 export { IStyle };
 
@@ -39,4 +20,5 @@ export type BasicFilters = BasicFilter[];
 export type DataCells = DataCell[];
 export type Cells = Cell[];
 export type Headers = Header[];
+
 export type Table = Style;